Rename home page content file for Hugo 0.32+ compatibility
Starting in Hugo v0.32, an index.md file will be treated as a page bundle,
meaning other files in the same folder will be treated as "resources" instead of
being built as separate pages.

Based on this forum post:
https://discourse.gohugo.io/t/solved-site-no-longer-builds-with-version-0-32-theme-hugo-foundation6/9944/6

Renaming the file to something that is not "index" or "_index" and then manually
setting the URL in the Frontmatter to "index.html" basically fixes the problem.

However, the Quire CLI also needs a small update to be aware of the frontmatter
URL property of a given page, in order to properly build PDF/Epub versions of
the same content.

Hopefully more documentation on Page Bundles will be available soon:
https://gohugo.io/about/new-in-032/